# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Scan -\u003e Plan -\u003e Approve -\u003e Fix -\u003e Rollback REST API for golangci-lint issue triage.",
        "title": "Golangci Lint Dashboard API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/config": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "History and config endpoints are M5+ scope -- see golangci/plans/15-implementation-plan.md.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unimplemented"
                ],
                "summary": "Not implemented yet",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/fixes": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rule BR-3: the plan must be \"approved\". Rule BR-1: one scan or fix at a time per repo+branch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fixes"
                ],
                "summary": "Apply an approved fix plan",
                "parameters": [
                    {
                        "description": "Approved plan to apply",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createFixRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "409": {
                        "description": "plan not approved, or a scan/fix already running for this repo+branch",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/fixes/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fixes"
                ],
                "summary": "Get a fix application's result by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fix history ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.FixHistory"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/history/fixes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "History and config endpoints are M5+ scope -- see golangci/plans/15-implementation-plan.md.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unimplemented"
                ],
                "summary": "Not implemented yet",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/history/rollbacks": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "History and config endpoints are M5+ scope -- see golangci/plans/15-implementation-plan.md.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unimplemented"
                ],
                "summary": "Not implemented yet",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/history/scans": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "History and config endpoints are M5+ scope -- see golangci/plans/15-implementation-plan.md.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Unimplemented"
                ],
                "summary": "Not implemented yet",
                "responses": {
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/issues/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Issues"
                ],
                "summary": "Get a lint issue by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Issue ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.LintIssue"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/plans": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Cache hit (same issue batch already planned) returns 200 with the ready plan immediately; a cache miss enqueues generation and returns 202.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plans"
                ],
                "summary": "Request an AI-generated fix plan",
                "parameters": [
                    {
                        "description": "Issue IDs to plan a fix for",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createPlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cache hit -- plan already ready",
                        "schema": {
                            "$ref": "#/definitions/storage.FixPlan"
                        }
                    },
                    "202": {
                        "description": "cache miss -- plan generation enqueued",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/plans/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plans"
                ],
                "summary": "Get a fix plan by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.FixPlan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/plans/{id}/approve": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rule BR-3: applying a fix requires an approved plan. Rule BR-5: high-risk/breaking-change plans require explicit confirmed=true.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plans"
                ],
                "summary": "Approve or reject a fix plan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "approve defaults to true; set false to reject",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.approvePlanRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.FixPlan"
                        }
                    },
                    "400": {
                        "description": "confirmation required for high-risk/breaking plans",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "409": {
                        "description": "plan is not pending an approval decision",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/rollbacks": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Rule BR-4: revert-only, never reset/force-push. Fails if the fix made no commit, or was already rolled back.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollbacks"
                ],
                "summary": "Roll back an applied fix",
                "parameters": [
                    {
                        "description": "Fix to roll back",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createRollbackRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "fix made no commit -- nothing to roll back",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "409": {
                        "description": "already rolled back, or a scan/fix already running for this repo+branch",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/rollbacks/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Rollbacks"
                ],
                "summary": "Get a rollback's result by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Rollback ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.RollbackHistory"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/scans": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Enqueues a golangci-lint scan for repo_ref+branch. Rule BR-1: only one scan or fix may run at a time per repo+branch.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scans"
                ],
                "summary": "Start a lint scan",
                "parameters": [
                    {
                        "description": "Scan target",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.createScanRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    },
                    "409": {
                        "description": "a scan or fix is already running for this repo+branch",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/scans/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scans"
                ],
                "summary": "Get a scan by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.LintScan"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        },
        "/scans/{id}/issues": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Scans"
                ],
                "summary": "List issues found by a scan",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Scan ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 50, capped at 200)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.LintIssue"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/golangci_backend_frontiir_utils.RestErr"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.approvePlanRequest": {
            "type": "object",
            "properties": {
                "approve": {
                    "description": "Approve defaults to true (JSON omitted -\u003e zero value -\u003e approve) --\nthe endpoint's own name is \"approve\"; explicit {\"approve\": false} is\nhow a reject happens (06-api-design.md has no separate reject route).",
                    "type": "boolean"
                },
                "confirmed": {
                    "type": "boolean"
                }
            }
        },
        "api.createFixRequest": {
            "type": "object",
            "required": [
                "plan_id"
            ],
            "properties": {
                "plan_id": {
                    "type": "string"
                }
            }
        },
        "api.createPlanRequest": {
            "type": "object",
            "required": [
                "issue_ids"
            ],
            "properties": {
                "issue_ids": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "api.createRollbackRequest": {
            "type": "object",
            "required": [
                "fix_history_id"
            ],
            "properties": {
                "fix_history_id": {
                    "type": "string"
                }
            }
        },
        "api.createScanRequest": {
            "type": "object",
            "required": [
                "branch",
                "repo_ref"
            ],
            "properties": {
                "branch": {
                    "type": "string"
                },
                "repo_ref": {
                    "type": "string"
                }
            }
        },
        "golangci_backend_frontiir_utils.RestErr": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Machine-readable business error code (e.g. \"SLA_EXPIRED\")",
                    "type": "string"
                },
                "data": {
                    "description": "Optional structured payload (e.g. conflicting ticket detail)"
                },
                "message": {
                    "description": "User-friendly error message",
                    "type": "string"
                },
                "status": {
                    "description": "HTTP status code",
                    "type": "integer"
                }
            }
        },
        "storage.FixHistory": {
            "type": "object",
            "properties": {
                "applied_by": {
                    "type": "string"
                },
                "branch_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "diff_ref": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "plan_id": {
                    "type": "string"
                },
                "post_fix_scan_id": {
                    "type": "string"
                },
                "pre_fix_scan_id": {
                    "type": "string"
                },
                "result": {
                    "description": "applying|passed|failed -- \"applying\" added in M4, apply is worker-queued like scan/plan",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "storage.LintIssue": {
            "type": "object",
            "properties": {
                "column": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "fingerprint": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "line": {
                    "type": "integer"
                },
                "linter": {
                    "type": "string"
                },
                "message": {
                    "description": "golangci-lint's issue text, minus the rule prefix -- see parser.Parse",
                    "type": "string"
                },
                "plans": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.FixPlan"
                    }
                },
                "reason_my": {
                    "description": "Burmese \"why this breaks the rule\" explanation -- see parser.reasonMy",
                    "type": "string"
                },
                "rule": {
                    "type": "string"
                },
                "scan_id": {
                    "type": "string"
                },
                "severity": {
                    "description": "critical|high|medium|low|info",
                    "type": "string"
                },
                "status": {
                    "description": "open|planned|fix_applied|resolved|reopened|ignored",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.LintScan": {
            "type": "object",
            "properties": {
                "branch": {
                    "type": "string"
                },
                "commit_sha": {
                    "type": "string"
                },
                "config_hash": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "raw_output_ref": {
                    "type": "string"
                },
                "repo_ref": {
                    "type": "string"
                },
                "status": {
                    "description": "running|success|failed",
                    "type": "string"
                },
                "triggered_by": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.RollbackHistory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "fix_history_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "post_rollback_scan_id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "result": {
                    "description": "reverting|done|conflict|failed -- added when POST /rollbacks was\nimplemented: 11-sequence-diagram.md's own Flow 4 describes two\noutcomes (revert succeeds vs conflicts) that need to be\ndistinguishable, and apply is worker-queued like scan/plan/fix so\nneeds an in-flight state too, but no such field existed.",
                    "type": "string"
                },
                "revert_commit_sha": {
                    "type": "string"
                },
                "rolled_back_by": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}